Refactor dg plus CLI into organized module structure
Breaks up 384-line monolithic __init__.py into organized modules:
- login.py for authentication commands
- pull.py for pull commands
- create/ module for creation commands (env, ci-api-token)
- deploy/ module for deployment commands
- Maintains backward compatibility with no functional changes
