+++ python_modules/libraries/dagster-dg-cli/dagster_dg_cli/cli/plus/deploy/__init__.py
@@ -1,0 +1,4 @@
+from dagster_dg_cli.cli.plus.deploy.commands import (
+    DEFAULT_STATEDIR_PATH as DEFAULT_STATEDIR_PATH,
+    deploy_group as deploy_group,
+)
