+++ python_modules/libraries/dagster-dg-cli/dagster_dg_cli/cli/plus/deploy/deploy_session.py
@@ -4,11 +4,16 @@ import subprocess
 import sys
 import uuid
 from pathlib import Path
-from typing import Optional
+from typing import TYPE_CHECKING, Optional
 
 import click
+
+if TYPE_CHECKING:
+    from dagster_cloud_cli.types import SnapshotBaseDeploymentCondition
+
 import dagster_shared.check as check
-from dagster_cloud_cli.types import SnapshotBaseDeploymentCondition
+
+# Expensive imports moved to lazy loading inside functions to improve CLI startup performance
 from dagster_dg_core.config import DgRawBuildConfig, merge_build_configs
 from dagster_dg_core.context import DgContext
 from dagster_dg_core.utils.git import get_local_branch_name
@@ -57,8 +62,6 @@ def _build_hybrid_image(
     merged_build_config: DgRawBuildConfig,
     workspace_context: Optional[DgContext],
 ) -> None:
-    from dagster_cloud_cli.commands.ci import set_build_output_impl
-
     registry = merged_build_config.get("registry")
 
     if not registry:
@@ -102,6 +105,9 @@ def _build_hybrid_image(
 
     subprocess.run(push_cmd, check=True)
 
+    # Lazy import for test mocking and performance
+    from dagster_cloud_cli.commands.ci import set_build_output_impl
+
     set_build_output_impl(
         statedir=str(statedir),
         location_name=[dg_context.code_location_name],
@@ -120,10 +126,8 @@ def init_deploy_session(
     commit_hash: Optional[str],
     location_names: tuple[str],
     status_url: Optional[str],
-    snapshot_base_condition: Optional[SnapshotBaseDeploymentCondition],
+    snapshot_base_condition: Optional["SnapshotBaseDeploymentCondition"],
 ):
-    from dagster_cloud_cli.commands.ci import init_impl
-
     deployment_type = (
         input_deployment_type
         if input_deployment_type
@@ -139,6 +143,9 @@ def init_deploy_session(
         os.makedirs(statedir)
 
     dagster_cloud_yaml_file = create_temp_dagster_cloud_yaml_file(dg_context, statedir)
+
+    # Lazy import for test mocking and performance
+    from dagster_cloud_cli.commands.ci import init_impl
 
     init_impl(
         statedir=str(statedir),
@@ -209,9 +216,6 @@ def _build_artifact_for_project(
     python_version: str,
     workspace_context: Optional[DgContext],
 ):
-    from dagster_cloud_cli.commands.ci import BuildStrategy, build_impl
-    from dagster_cloud_cli.core.pex_builder import deps
-
     merged_build_config: DgRawBuildConfig = merge_build_configs(
         workspace_context.build_config if workspace_context else None,
         dg_context.build_config,
@@ -241,6 +245,11 @@ def _build_artifact_for_project(
         )
 
     else:
+        # Import BuildStrategy and deps locally since they're not needed for tests
+        # Lazy import for test mocking and performance
+        from dagster_cloud_cli.commands.ci import BuildStrategy, build_impl
+        from dagster_cloud_cli.core.pex_builder import deps
+
         build_impl(
             statedir=str(statedir),
             dockerfile_path=str(dockerfile_path),
@@ -260,6 +269,7 @@ def _build_artifact_for_project(
 
 
 def finish_deploy_session(dg_context: DgContext, statedir: str, location_names: tuple[str]):
+    # Lazy imports for test mocking and performance
     from dagster_cloud_cli.commands.ci import deploy_impl
     from dagster_cloud_cli.config_utils import (
         get_agent_heartbeat_timeout,
