+++ python_modules/libraries/dagster-dg-cli/tox.ini
@@ -15,6 +15,8 @@ install_command =
 
 deps =
   pytest
+  typer
+  mcp; python_version>='3.10'
   -e ../dagster-dg-core[test]
   -e ../../../python_modules/dagster[test]
   -e ../../../python_modules/dagster-test
